+++ javafx-ui-controls/src/com/sun/javafx/scene/control/skin/RadioButtonSkin.java
@@ -77,11 +77,11 @@ public class RadioButtonSkin extends LabeledSkinBase<RadioButton, ButtonBehavior
      *                                                                         *
      **************************************************************************/
     
-    @Override protected double computeMinWidth(double height, int topInset, int rightInset, int bottomInset, int leftInset) {
+    @Override protected double computeMinWidth(double height, double topInset, double rightInset, double bottomInset, double leftInset) {
         return super.computePrefWidth(height, topInset, rightInset, bottomInset, leftInset) + snapSize(radio.minWidth(-1));
     }
     
-    @Override protected double computeMinHeight(double width, int topInset, int rightInset, int bottomInset, int leftInset) {
+    @Override protected double computeMinHeight(double width, double topInset, double rightInset, double bottomInset, double leftInset) {
         return Math.max(snapSize(super.computeMinHeight(width - radio.minWidth(-1), topInset, rightInset, bottomInset, leftInset)),
                 topInset + radio.minHeight(-1) + bottomInset);
     }
